backend/requirements_dev.txt
Django==3.2.6
gunicorn==20.1.0
psycopg2-binary==2.9.1
Pillow
django-ckeditor==6.5.1
djangorestframework
django-cors-headers
drf-spectacular